source_codes/gralaf/requirements-gralaf.txt
causalnex==0.11.1
networkx==2.8.5
numpy==1.22.4
pandas==1.4.3
pygraphviz==1.10
PyYAML==6.0
requests==2.28.1
scikit-learn==1.2.2
matplotlib

